FAR patch 00135.warnings Дата              : 07.07.2006 Сделал            : Ivan Sintyurin Описание          : - warnings Измененные файлы  : cddrv.cpp checkver.cpp clipboard.cpp cmdline.cpp copy.cpp                     delete.cpp dialog.cpp dizlist.cpp dizlist.hpp edit.cpp                     editor.cpp eject.cpp execute.cpp farconst.hpp fileedit.cpp                     filelist.cpp fileview.cpp filter.hpp findfile.cpp                     flink.cpp flshow.cpp flupdate.cpp fnparce.cpp help.cpp                     hilight.cpp hotplug.cpp infolist.cpp local.cpp macro.cpp                     message.cpp mix.cpp panel.cpp plugapi.cpp plugins.cpp                     strmix.cpp struct.hpp syntax.cpp syslog.cpp treelist.cpp                     UnicodeString.hpp viewer.cpp viewer.hpp xlat.cpp Новые файлы       : Удаленные файлы   : Состав            : 00135.warnings.txt                     cddrv.cpp.135.diff          checkver.cpp.135.diff                     clipboard.cpp.135.diff      cmdline.cpp.135.diff                     copy.cpp.135.diff           delete.cpp.135.diff                     dialog.cpp.135.diff         dizlist.cpp.135.diff                     dizlist.hpp.135.diff        edit.cpp.135.diff                     editor.cpp.135.diff         eject.cpp.135.diff                     execute.cpp.135.diff        farconst.hpp.135.diff                     fileedit.cpp.135.diff       filelist.cpp.135.diff                     fileview.cpp.135.diff       filter.hpp.135.diff                     findfile.cpp.135.diff       flink.cpp.135.diff                     flshow.cpp.135.diff         flupdate.cpp.135.diff                     fnparce.cpp.135.diff        help.cpp.135.diff                     hilight.cpp.135.diff        hotplug.cpp.135.diff                     infolist.cpp.135.diff       local.cpp.135.diff                     macro.cpp.135.diff          message.cpp.135.diff                     mix.cpp.135.diff            panel.cpp.135.diff                     plugapi.cpp.135.diff        plugins.cpp.135.diff                     strmix.cpp.135.diff         struct.hpp.135.diff                     syntax.cpp.135.diff         syslog.cpp.135.diff                     treelist.cpp.135.diff       UnicodeString.hpp.135.diff                     viewer.cpp.135.diff         viewer.hpp.135.diff                     xlat.cpp.135.diff Основан на патче  : 134 Дополнение        :
Как и в случае с 1.7х остались предупреждения, касающиеся SEH и деструкторов.

Попутно исправлены баги:
 - в dialog.cpp, hotplug.cpp, plugapi.cpp перепутаны параметры конструктора
   VMenu
 - кое-где (видимо, при внедрении юникода) перепутаны были char и wchar_t

Просьба остается прежней - писать без warnings.

+++ far/copy.cpp
@@ -5,7 +5,7 @@ copy.cpp
 
 */
 
-/* Revision: 1.186 26.06.2006 $ */
+/* Revision: 1.187 07.07.2006 $ */
 
 #include "headers.hpp"
 #pragma hdrstop
@@ -3372,9 +3372,9 @@ int ShellCopy::ShowBar(unsigned __int64 WrittenSize,unsigned __int64 TotalSize,b
     Length=BarLength;
   else
     if (TotalSize<1000000)
-      Length=WrittenSize*BarLength/TotalSize;
+      Length=static_cast<int>(WrittenSize*BarLength/TotalSize);
     else
-      Length=(WrittenSize/100)*BarLength/(TotalSize/100);
+      Length=static_cast<int>((WrittenSize/100)*BarLength/(TotalSize/100));
   wchar_t ProgressBar[100];
 
   for (int i = 0; i < BarLength; i++)
@@ -3423,8 +3423,8 @@ int ShellCopy::ShowBar(unsigned __int64 WrittenSize,unsigned __int64 TotalSize,b
     {
       if (TotalBar)
         OldWrittenSize = OldWrittenSize - TotalSkippedSize;
-      int CPS = OldWrittenSize/WorkTime;
-      TimeLeft = (CPS)?SizeLeft/CPS:0;
+      int CPS = static_cast<int>(OldWrittenSize/WorkTime);
+      TimeLeft = static_cast<int>((CPS)?SizeLeft/CPS:0);
       c[0]=' ';
       if (CPS > 99999) {
         c[0]='K';
@@ -3708,7 +3708,7 @@ BOOL ShellCopySecuryMsgW(const wchar_t *Name)
 
   string strOutFileName;
 
-  if (Name == NULL || *Name == 0 || ((clock() - PrepareSecuryStartTime) > Opt.ShowTimeoutDACLFiles))
+  if (Name == NULL || *Name == 0 || (static_cast<DWORD>(clock() - PrepareSecuryStartTime) > Opt.ShowTimeoutDACLFiles))
   {
     if(Name && *Name)
     {
